pyproject.toml
[project]
name = "kg-eval"
version = "0.1.0"
description = "A comprehensive framework for evaluating Large Language Models' knowledge graph construction capabilities"
readme = "README.md"
authors = [
    { name = "earayu", email = "earayu@163.com" }
]
requires-python = ">=3.11"
dependencies = [
    "networkx>=3.0",
    "numpy>=1.20.0",
    "pandas>=1.3.0",
    "scikit-learn>=1.0.0",
    "matplotlib>=3.5.0",
    "plotly>=5.0.0",
    "pydantic>=2.0.0",
    "openai>=1.0.0",
    "anthropic>=0.7.0",
    "Levenshtein>=0.20.0",
    "tqdm>=4.64.0",
    "click>=8.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.scripts]
kg-eval = "kg_eval:main"

[project.urls]
Homepage = "https://github.com/earayu/KG-Eval"
Repository = "https://github.com/earayu/KG-Eval"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
